feat: add video timeline/scrubber
