+++ src-tauri/src/native_video/constants.rs
@@ -1,0 +1,16 @@
+pub(crate) const RESOURCE_SAMPLE_MS: u64 = 150;
+pub(crate) const BROKER_QUEUE_CAPACITY: usize = 96;
+pub(crate) const FRAME_BROADCAST_CAPACITY: usize = 32;
+pub(crate) const FRAME_PACKET_HEADER_LEN: usize = 64;
+pub(crate) const FRAME_PACKET_MAGIC: &[u8; 4] = b"SVF1";
+pub(crate) const BYTES_PER_PIXEL_RGBA8: u64 = 4;
+pub(crate) const SAFE_BUDGET_FACTOR: f64 = 0.8;
+pub(crate) const BASE_CASE_MAX_STREAMS_BEFORE_VALIDATION: usize = 1;
+pub(crate) const SCALING_MAX_STREAMS_AFTER_VALIDATION: usize = 32;
+pub(crate) const DOWNGRADE_QUEUE_PRESSURE: f64 = 0.65;
+pub(crate) const DOWNGRADE_DROP_RATE: f64 = 0.03;
+pub(crate) const UPGRADE_QUEUE_PRESSURE: f64 = 0.25;
+pub(crate) const UPGRADE_HEADROOM: f64 = 0.25;
+pub(crate) const MIN_UPGRADE_DWELL_MS: u64 = 2_000;
+pub(crate) const MIN_DOWNGRADE_DWELL_MS: u64 = 800;
+pub(crate) const MATERIAL_OVERSAMPLE: f64 = 1.15;
